refactor: split stripe webhook route into handler modules
The route file carried signature verification, idempotency/ordering
guards, six event handlers, invoice-subscription resolution, and two
email senders in one 605-line file.

Extract everything but request handling into src/features/payments/webhooks/,
where vitest already collects tests (the suite only includes
src/features/** and src/shared/**), leaving route.ts at 134 lines doing
verification, guards, and dispatch.

Pure extraction — no behavior change. Handlers import the same module
specifiers as before, so existing test mocks apply unchanged.

Verified: 344/344 tests pass; tsc error signature identical to baseline;
eslint clean on all touched files.
